wait to confirm split
