+++ hw5/eesss34690/matrix.hpp
@@ -1,0 +1,214 @@
+#include <cstddef>
+#include <vector>
+#include <memory>
+#include <cstdlib>
+#include <new>
+#include <iostream>
+
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
+#include <pybind11/operators.h>
+
+#include <algorithm>
+#include <atomic>
+#include "mkl.h"
+
+namespace py = pybind11;
+using namespace std;
+
+struct ByteCounterImpl
+{
+
+    std::size_t allocated = 0;
+    std::size_t deallocated = 0;
+    std::size_t refcount = 0;
+
+}; /* end struct ByteCounterImpl */
+
+/**
+ * One instance of this counter is shared among a set of allocators.
+ *
+ * The counter keeps track of the bytes allocated and deallocated, and report
+ * those two numbers in addition to bytes that remain allocated.
+ */
+class ByteCounter
+{
+
+public:
+
+    ByteCounter()
+      : m_impl(new ByteCounterImpl)
+    { incref(); }
+
+    ByteCounter(ByteCounter const & other)
+      : m_impl(other.m_impl)
+    { incref(); }
+
+    ByteCounter & operator=(ByteCounter const & other)
+    {
+        if (&other != this)
+        {
+            decref();
+            m_impl = other.m_impl;
+            incref();
+        }
+
+        return *this;
+    }
+
+    ByteCounter(ByteCounter && other)
+      : m_impl(other.m_impl)
+    { incref(); }
+
+    ByteCounter & operator=(ByteCounter && other)
+    {
+        if (&other != this)
+        {
+            decref();
+            m_impl = other.m_impl;
+            incref();
+        }
+
+        return *this;
+    }
+
+    ~ByteCounter() { decref(); }
+
+    void swap(ByteCounter & other)
+    {
+        std::swap(m_impl, other.m_impl);
+    }
+
+    void increase(std::size_t amount)
+    {
+        m_impl->allocated += amount;
+    }
+
+    void decrease(std::size_t amount)
+    {
+        m_impl->deallocated += amount;
+    }
+
+    std::size_t bytes() const { return m_impl->allocated - m_impl->deallocated; }
+    std::size_t allocated() const { return m_impl->allocated; }
+    std::size_t deallocated() const { return m_impl->deallocated; }
+    /* This is for debugging. */
+    std::size_t refcount() const { return m_impl->refcount; }
+
+private:
+
+    void incref() { ++m_impl->refcount; }
+
+    void decref()
+    {
+        if (nullptr == m_impl)
+        {
+            // Do nothing.
+        }
+        else if (1 == m_impl->refcount)
+        {
+            delete m_impl;
+            m_impl = nullptr;
+        }
+        else
+        {
+            --m_impl->refcount;
+        }
+    }
+
+    ByteCounterImpl * m_impl;
+
+}; /* end class ByteCounter */
+
+template <class T>
+struct MyAllocator
+{
+
+    using value_type = T;
+
+    // Just use the default constructor of ByteCounter for the data member
+    // "counter".
+    MyAllocator() = default;
+
+    template <class U> constexpr
+    MyAllocator(const MyAllocator<U> & other) noexcept
+    {
+        counter = other.counter;
+    }
+
+    T * allocate(std::size_t n)
+    {
+        if (n > std::numeric_limits<std::size_t>::max() / sizeof(T))
+        {
+            throw std::bad_alloc();
+        }
+        const std::size_t bytes = n*sizeof(T);
+        T * p = static_cast<T *>(std::malloc(bytes));
+        if (p)
+        {
+            counter.increase(bytes);
+            return p;
+        }
+        else
+        {
+            throw std::bad_alloc();
+        }
+    }
+
+    void deallocate(T* p, std::size_t n) noexcept
+    {
+        std::free(p);
+
+        const std::size_t bytes = n*sizeof(T);
+        counter.decrease(bytes);
+    }
+
+    ByteCounter counter;
+
+}; /* end struct MyAllocator */
+
+template <class T, class U>
+bool operator==(const MyAllocator<T> & a, const MyAllocator<U> & b)
+{
+    return a.counter == b.counter;
+}
+
+template <class T, class U>
+bool operator!=(const MyAllocator<T> & a, const MyAllocator<U> & b)
+{
+    return !(a == b);
+}
+
+static MyAllocator<double> alloc;
+
+class Matrix {
+ public:
+  Matrix(size_t nrow, size_t ncol)
+      : m_nrow(nrow), m_ncol(ncol), m_buffer(alloc)
+  {
+  	m_buffer.resize(m_nrow* m_ncol, 0.0);
+  }
+  Matrix(const vector<vector<double>> &m);
+
+  bool operator==(const Matrix &other);
+  double operator()(size_t row, size_t col) const {
+    return m_buffer[row * m_ncol + col];
+  }
+  double &operator()(size_t row, size_t col) {
+    return m_buffer[row * m_ncol + col];
+  }
+
+  size_t nrow() const { return m_nrow; }
+  size_t ncol() const { return m_ncol; }
+  double *buffer() { return m_buffer.data(); }
+  const double *buffer() const { return m_buffer.data(); }
+
+ private:
+  size_t m_nrow;
+  size_t m_ncol;
+  vector<double, MyAllocator<double>> m_buffer;
+};
+
+size_t bytes() { return alloc.counter.bytes(); }
+size_t allocated() { return alloc.counter.allocated(); }
+size_t deallocated() { return alloc.counter.deallocated(); }
